:recycle: 인덱스 -> 홈으로 변경하고 배너 컨트롤러 분리
