+++ src/main/java/container/restaurant/server/web/HomeController.java
@@ -1,12 +1,11 @@
 package container.restaurant.server.web;
 
 import container.restaurant.server.config.auth.LoginId;
-import container.restaurant.server.domain.home.banner.BannerService;
 import container.restaurant.server.domain.home.phrase.PhraseService;
 import container.restaurant.server.domain.statistics.StatisticsService;
 import container.restaurant.server.domain.user.User;
 import container.restaurant.server.domain.user.UserService;
-import container.restaurant.server.web.dto.IndexDto;
+import container.restaurant.server.web.dto.HomeDto;
 import container.restaurant.server.web.linker.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.hateoas.Link;
@@ -21,43 +20,42 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/")
-public class IndexController {
+@RequestMapping("api/home")
+public class HomeController {
 
     private final UserLinker userLinker;
     private final FeedLinker feedLinker;
     private final StatisticsLinker statisticsLinker;
-    private final IndexLinker indexLinker;
+    private final BannerLinker bannerLinker;
     private final RestaurantLinker restaurantLinker;
 
     private final UserService userService;
     private final StatisticsService statisticsService;
-    private final BannerService bannerService;
     private final PhraseService phraseService;
 
     @GetMapping
-    public ResponseEntity<?> index(@LoginId Long loginId) {
+    public ResponseEntity<?> home(@LoginId Long loginId) {
         User loginUser = loginId != null ? userService.findById(loginId) : null;
 
         return ResponseEntity.ok(
-                setLinks(IndexDto.builder()
+                setLinks(HomeDto.builder()
                         .user(loginUser)
                         .totalContainer(statisticsService.getTotalFeed())
                         .phrase(phraseService.getPhrase())
                         .build()));
     }
 
-    private IndexDto setLinks(IndexDto model) {
+    private HomeDto setLinks(HomeDto model) {
         Long loginId = model.getLoginId();
         return model
                 .add(List.of(
-                        linkTo(IndexController.class).withSelfRel(),
+                        linkTo(HomeController.class).withSelfRel(),
                         feedLinker.selectFeed().withRel("feed-list"),
                         feedLinker.selectRecommend().withRel("feed-recommend"),
                         feedLinker.createFeed().withRel("feed-create"),
                         statisticsLinker.getFeedCountTopUsers().withRel("top-users"),
                         statisticsLinker.getRecentFeedUsers().withRel("recent-users"),
-                        indexLinker.getBanners().withRel("banner-list"),
+                        bannerLinker.getBanners().withRel("banner-list"),
                         restaurantLinker.findNearByRestaurants().withRel("restaurant-near")
                 ))
                 .addAllIf(loginId == null, () -> List.of(
@@ -67,13 +65,6 @@ public class IndexController {
                         Link.of("/logout").withRel("logout"),
                         userLinker.getCurrentUser().withRel("my-info")
                 ));
-    }
-
-    @GetMapping("/banners")
-    public ResponseEntity<?> getBanners(){
-        return ResponseEntity.ok(
-                bannerService.getBanners()
-        );
     }
 
 }
