+++ src/main/java/container/restaurant/server/web/linker/BannerLinker.java
@@ -1,6 +1,6 @@
 package container.restaurant.server.web.linker;
 
-import container.restaurant.server.web.IndexController;
+import container.restaurant.server.web.BannerController;
 import org.springframework.hateoas.server.LinkBuilder;
 import org.springframework.hateoas.server.core.DummyInvocationUtils;
 import org.springframework.stereotype.Component;
@@ -8,9 +8,11 @@ import org.springframework.stereotype.Component;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 
 @Component
-public class IndexLinker {
-    IndexController proxy
-            = DummyInvocationUtils.methodOn(IndexController.class);
+public class BannerLinker {
+
+    BannerController proxy
+            = DummyInvocationUtils.methodOn(BannerController.class);
 
     public LinkBuilder getBanners(){ return linkTo(proxy.getBanners());}
+
 }
