--- src/test/java/container/restaurant/server/web/IndexControllerTest.java
@@ -1,121 +1,0 @@
-package container.restaurant.server.web;
-
-import container.restaurant.server.domain.home.banner.Banner;
-import container.restaurant.server.domain.home.banner.BannerRepository;
-import container.restaurant.server.domain.home.banner.BannerService;
-import container.restaurant.server.web.base.BaseUserControllerTest;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.ResultActions;
-
-import static org.springframework.restdocs.hypermedia.HypermediaDocumentation.linkWithRel;
-import static org.springframework.restdocs.hypermedia.HypermediaDocumentation.links;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-class IndexControllerTest extends BaseUserControllerTest {
-
-    @Autowired
-    BannerRepository bannerRepository;
-
-    @Autowired
-    BannerService bannerService;
-
-    @Test
-    @DisplayName("비로그인 index 링크 테스트")
-    void testGuestIndexLinks() throws Exception {
-        String testPath = "/";
-
-        mvc.perform(get(testPath))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("_links.self.href").value(Matchers.containsString(testPath)))
-                .andExpect(jsonPath("_links.login.href").exists())
-                .andDo(document("index-guest",
-                        links(
-                                linkWithRel("self").description("본 응답의 링크"),
-                                linkWithRel("feed-list").description("전체 피드 리스트 링크 (페이징)"),
-                                linkWithRel("feed-recommend").description("추천 피드 리스트 링크"),
-                                linkWithRel("feed-create").description("피드를 생성하기 위한 링크"),
-                                linkWithRel("top-users").description("누적 피드가 최근 30일간 제일 많은 10명에 대한 리스트 링크"),
-                                linkWithRel("recent-users").description("최근에 피드를 작성한 사용자 리스트 링크"),
-                                linkWithRel("restaurant-near").description("주위 식당을 조회하기 위한 링크"),
-                                linkWithRel("banner-list").description("배너 리스트 링크"),
-                                linkWithRel("login").description("인증을 위한 링크 리스트")
-                        )
-        ));
-    }
-
-    @Test
-    @DisplayName("로그인 index 링크 테스트")
-    void testAuthIndexLinks() throws Exception {
-        String testPath = "/";
-
-        mvc.perform(
-                get(testPath)
-                        .session(myselfSession))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("_links.self.href").value(Matchers.containsString(testPath)))
-                .andExpect(jsonPath("_links.my-info").exists())
-                .andExpect(jsonPath("_links.logout.href").exists())
-                .andDo(document("index-user",
-                        links(
-                                linkWithRel("self").description("본 응답의 링크"),
-                                linkWithRel("feed-list").description("전체 피드 리스트 링크 (페이징)"),
-                                linkWithRel("feed-recommend").description("추천 피드 리스트 링크"),
-                                linkWithRel("feed-create").description("피드를 생성하기 위한 링크"),
-                                linkWithRel("top-users").description("누적 피드가 최근 30일간 제일 많은 10명에 대한 리스트 링크"),
-                                linkWithRel("recent-users").description("최근에 피드를 작성한 사용자 리스트 링크"),
-                                linkWithRel("restaurant-near").description("주위 식당을 조회하기 위한 링크"),
-                                linkWithRel("banner-list").description("배너 리스트 링크"),
-                                linkWithRel("my-info").description("사용자 상세 정보 링크"),
-                                linkWithRel("logout").description("현재 로그인을 로그아웃 하기 위한 링크")
-                        ),
-                        responseFields(
-                                fieldWithPath("loginId").description("로그인한 사용자 ID 비로그인인 경우 null"),
-                                fieldWithPath("myContainer").description("로그인한 사용자가 작성한 용기내 피드 개수"),
-                                fieldWithPath("totalContainer").description("어제까지 작성된 전체 용기내 피드 개수"),
-                                fieldWithPath("myLevelTitle").description("로그인한 사용자의 레벨 타이틀"),
-                                fieldWithPath("phrase").description("메인 문구"),
-                                subsectionWithPath("_links").description("본 응답에서 전이 가능한 링크 명세")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("배너 테스트")
-    public void getBanners() throws Exception {
-        for (int i = 0; i < 4; i++) {
-            bannerRepository.save(Banner.builder()
-                    .bannerURL("bannerURL" + i)
-                    .additionalURL("additionalURL" + i)
-                    .contentURL("contentURL" + i)
-                    .title("title" + i)
-                    .build());
-        }
-        bannerService.putBanners();
-
-
-        ResultActions perform = mvc.perform(
-                get("/banners")
-        );
-
-        perform
-                .andExpect(status().isOk())
-                .andDo(document("banners"))
-                .andDo(document("banner",
-                        responseFields(beneathPath("_embedded.bannerInfoDtoList"),
-                                fieldWithPath("title").description("배너 컨텐츠 이름"),
-                                fieldWithPath("bannerURL").description("메인 화면에서 보여질 이미지 URL"),
-                                fieldWithPath("contentURL").description("메인 화면에서 배너를 터치한 경우 보여질 이미지 URL"),
-                                fieldWithPath("additionalURL").description("배너 이미지 외에 추가로 제공해야할 URL 이 있는 경우 포함되는 URL")
-                        )));
-    }
-
-}
